Implements MapStruct mappers and enhances DTO validation
Introduces MapStruct-based mappers for all major entities to automate and standardize mapping between entities and DTOs, improving maintainability and reducing boilerplate. Strengthens validation on request DTOs to enforce data integrity, and refactors DTO field types for consistency. Updates Maven configuration to manage annotation processors and dependency versions, enabling annotation-driven code generation.

These changes lay the foundation for a cleaner, more robust service and API layer separation.

+++ src/main/java/com/portfolio/BiblioHub/book/dto/BookResponseDto.java
@@ -1,10 +1,6 @@
 package com.portfolio.BiblioHub.book.dto;
 
-import com.portfolio.BiblioHub.author.dto.AuthorResponseDto;
-import com.portfolio.BiblioHub.category.dto.CategoryResponseDto;
-import com.portfolio.BiblioHub.publisher.dto.PublisherResponseDto;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -25,9 +21,9 @@ public class BookResponseDto {
     private String publisherName;
 
     // Category details
-    private Integer categoryId;
+    private Long categoryId;       // better use Long instead of Integer
     private String categoryName;
 
-    // Author details (mini DTOs or just names)
+    // Authors (just names or mini DTOs)
     private List<String> authors;
 }
